Don't set _PACK_GROUP0 variable on non het-step
The commit applies the same logic to het-steps as we use for het jobs.
If we really have a heterogenou step we set environemnt variables with
_GROUP_PACKX suffix with X being het step offset. However, if we have
only one step (non het-step) we use the variable without the suffix.
This way we'll appropriately override varialbles like SLURM_MEM_PER_NODE
if we call salloc->srun->salloc multiple times with different --mem
requetes.

Bug 16240